requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
gunicorn==20.1.0
torch==2.0.0
torchvision==0.15.1
numpy==1.24.3
pillow==9.5.0
scipy==1.10.1
tqdm==4.65.0
jinja2==3.1.2